refactor(chat): refactor tab content components
